adds validator

+++ src/main/java/pt/isel/ls/handlers/misc/CloseHandler.java
@@ -1,9 +1,12 @@
 package pt.isel.ls.handlers.misc;
 
 import pt.isel.ls.app.http.HttpPool;
+import pt.isel.ls.exceptions.parameter.ValidatorException;
 import pt.isel.ls.handlers.RouteHandler;
 import pt.isel.ls.router.request.Method;
 import pt.isel.ls.router.request.RouteRequest;
+import pt.isel.ls.router.request.parameter.Validator;
+import pt.isel.ls.router.request.parameter.ValidatorResult;
 import pt.isel.ls.router.response.HandlerResponse;
 import pt.isel.ls.view.MessageView;
 
@@ -23,7 +26,15 @@ public class CloseHandler extends RouteHandler {
 
     @Override
     public HandlerResponse execute(RouteRequest request) {
-        int port = request.getParameter("port").get(0).toInt();
+        Validator validator = new Validator()
+                .addRule("port", p -> p.getUnique().toInt());
+
+        ValidatorResult res = validator.validate(request);
+        if (res.hasErrors()) {
+            throw new ValidatorException(res);
+        }
+
+        int port = res.getParameterValue("port");
         httpPool.terminate(port);
         return new HandlerResponse(new MessageView("HTTP Server on port " + port + " terminated!"));
     }
